Update belgium.ics
Created new holidays for the four coming years in Belgium. Recurring holidays are at the top of the document and variable date holidays are at the bottom, both in ascending order. Also, changed the holiday names to English instead of Flemish. We have three national languages (French, Flemish and German), and writing holidays in all three would be too long and not very useful. For that reason, English is often used as the "neutral" language between us.
(This is my first time editing, apologies if I have missed anything.)